LPD-18860 Implement card visualization mode field assignment tests
