--- modules/test/playwright/tests/frontend-data-set-views-web/fixtures/listVisualizationModePageTest.ts
@@ -1,18 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {test} from '@playwright/test';
-
-import {ListVisualizationModePage} from '../pages/view/visualization_modes/list/ListVisualizationModePage';
-
-const listVisualizationModePageTest = test.extend<{
-	listVisualizationModePage: ListVisualizationModePage;
-}>({
-	listVisualizationModePage: async ({page}, use) => {
-		await use(new ListVisualizationModePage(page));
-	},
-});
-
-export {listVisualizationModePageTest};
