+++ modules/test/playwright/tests/frontend-data-set-views-web/fixtures/visualizationModesPageTest.ts
@@ -1,0 +1,18 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {test} from '@playwright/test';
+
+import {VisualizationModesPage} from '../pages/view/visualization_modes/VisualizationModesPage';
+
+const visualizationModesPageTest = test.extend<{
+	visualizationModesPage: VisualizationModesPage;
+}>({
+	visualizationModesPage: async ({page}, use) => {
+		await use(new VisualizationModesPage(page));
+	},
+});
+
+export {visualizationModesPageTest};
